requirements.txt
fastapi
uvicorn
gradio
anthropic
python-dotenv
sqlalchemy[asyncio]
asyncpg